types

+++ keyhandler.h
@@ -11,7 +11,7 @@
 #define SCREEN_COLS 64
 #define TYPE_ROW_MIN 21
 
-void update_pressed(int*, int*, int*);
+void update_pressed(int*, uint8_t*, uint8_t*);
 void update_position(int*, int, char*, int*, int*);
 void parse_letters(int, int, int, char*, int*, int*);
 void parse_entry(int, int, int, char*, int*, int*);
